refactor: 사용자 로직 개선 - 로그 추가 - JWT 로직을 위한 필드 추가
