+++ webtoon/src/main/kotlin/com/rlc/webtoon/controller/UserController.kt
@@ -1,8 +1,10 @@
 package com.rlc.webtoon.controller
 
+import com.rlc.webtoon.dto.common.ResponseResult
 import com.rlc.webtoon.dto.request.UserRequestDto
 import com.rlc.webtoon.dto.response.UserResponseDto
 import com.rlc.webtoon.service.UserService
+import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -14,12 +16,15 @@ class UserController(
     private val userService: UserService
 ) {
 
+    private val log = LoggerFactory.getLogger(this::class.java)
+
     /**
      * 사용자 등록.
-     * TODO 응답 포맷 설정, 테스트 코드
      */
-    @PostMapping
-    fun signupUser(@RequestBody userRequestDto: UserRequestDto): UserResponseDto {
-        return userService.signupUser(userRequestDto)
+    @PostMapping("/signup")
+    fun signupUser(@RequestBody userRequestDto: UserRequestDto): ResponseResult<UserResponseDto> {
+        log.info("signupUser(), userRequestDto:$userRequestDto")
+
+        return ResponseResult(userService.signupUser(userRequestDto))
     }
 }
