+++ webtoon/src/main/kotlin/com/rlc/webtoon/dto/response/UserResponseDto.kt
@@ -4,6 +4,7 @@ import com.rlc.webtoon.entity.User
 import java.time.Instant
 
 data class UserResponseDto constructor(
+    val uuid: String,
     val id: String,
     val email: String,
     val phoneNumber: String,
@@ -14,7 +15,8 @@ data class UserResponseDto constructor(
     companion object {
         fun of(user: User): UserResponseDto {
             return UserResponseDto(
-                id = user.id,
+                uuid = user.uuid!!,
+                id = user.loginId,
                 email = user.email,
                 phoneNumber = user.phoneNumber,
                 birth = user.birth,
